+++ components/TrendingCard.tsx
@@ -4,15 +4,16 @@ import { Link } from "expo-router";
 import React from "react";
 import { Image, Text, TouchableOpacity, View } from "react-native";
 
-const TrendingCard = ({
-  movie: { movie_id, title, poster_url },
-  index,
-}: TrendingCardProps) => {
+const TrendingCard = ({ movie: { id, title, poster_path }, index }: any) => {
   return (
-    <Link href={`/movies/${movie_id}`} asChild>
+    <Link href={`/movies/${id}`} asChild>
       <TouchableOpacity className="w-32 relative pl-5">
         <Image
-          source={{ uri: poster_url }}
+          source={{
+            uri: poster_path
+              ? `https://image.tmdb.org/t/p/w500${poster_path}`
+              : "https://placehold.co/600x400/1a1a1a/ffffff.png",
+          }}
           className="w-32 h-48 rounded-2xl"
           resizeMode="cover"
         />